+++ nexign-kotlin-bp-dsl-engine/src/main/kotlin/com/nexign/dsl/engine/worker/Worker.kt
@@ -1,8 +1,8 @@
 package com.nexign.dsl.engine.worker
 
 import com.nexign.dsl.base.Operation
-import com.nexign.dsl.base.STOP_EXECUTION
 import com.nexign.dsl.base.Scenario
+import com.nexign.dsl.base.transitions.STOP_EXECUTION
 
 class Worker {
 
@@ -19,7 +19,7 @@ class Worker {
             if (condition == STOP_EXECUTION) {
                 break
             }
-            val nextOp = scenario.specification[currentOp]?.get(condition)
+            val nextOp = scenario.specification.routing[currentOp]?.get(condition)
                 ?: throw IllegalStateException() // TODO: change to custom
             currentOp = nextOp
         }
