+++ nexign-kotlin-bp-dsl-scenarios/src/main/kotlin/com/nexign/dsl/scenarios/examples/arithmeticscenario/ArithmeticScenario.kt
@@ -1,15 +1,19 @@
 package com.nexign.dsl.scenarios.examples.arithmeticscenario
 
 import com.nexign.dsl.base.Scenario
-import com.nexign.dsl.base.Specification
-import com.nexign.dsl.base.specification
+import com.nexign.dsl.base.specification.Specification
+import com.nexign.dsl.base.specification.routing
+import com.nexign.dsl.base.specification.specification
 
 class ArithmeticScenario(store: MutableMap<String, Any>) : Scenario(store) {
 
     override val specification: Specification = specification {
-        start(ValidateOr() binary {
-            yes(ComputeSquare() next ComputePerimeter() next PrintResults())
-            no(PrintError())
-        })
+        routing = routing {
+            start(ValidateOr() binary {
+                yes(ComputeSquare() next ComputePerimeter() next PrintResults())
+                no(PrintError())
+
+            })
+        }
     }
 }
