+++ nexign-kotlin-bp-dsl-scenarios/src/main/kotlin/com/nexign/dsl/scenarios/examples/bpscenario/ExampleOperations.kt
@@ -1,38 +1,47 @@
 package com.nexign.dsl.scenarios.examples.bpscenario
 
 import com.nexign.dsl.base.*
+import com.nexign.dsl.base.transitions.*
 
 import kotlin.random.Random
 
 open class GetAbonentInfo : Operation() {
 
     override val func: Scenario.() -> TransitionCondition = {
-        // Do something
-        SINGLE_ROUTE
+        this.functionBuilder {
+            // Do something
+            SINGLE_ROUTE
+        }
     }
 }
 
 open class ProlongAction : Operation() {
 
     override val func: Scenario.() -> TransitionCondition = {
-        // Do something
-        SINGLE_ROUTE
+        this.functionBuilder {
+            // Do something
+            SINGLE_ROUTE
+        }
     }
 }
 
 open class ActivateAction : Operation() {
 
     override val func: Scenario.() -> TransitionCondition = {
-        // Do something
-        SINGLE_ROUTE
+        this.functionBuilder {
+            // Do something
+            SINGLE_ROUTE
+        }
     }
 }
 
 open class CancelActionActivation : Operation() {
 
     override val func: Scenario.() -> TransitionCondition = {
-        // Do something
-        SINGLE_ROUTE
+        this.functionBuilder {
+            // Do something
+            SINGLE_ROUTE
+        }
     }
 }
 
@@ -42,39 +51,67 @@ open class NotifyAction(
 ) : Operation() {
 
     override val func: Scenario.() -> TransitionCondition = {
-        // Do something
-        println("I am notified about $message")
+        this.functionBuilder {
+            // Do something
+            println("I am notified about $message")
 
-        SINGLE_ROUTE
+            SINGLE_ROUTE
+        }
     }
 }
 
 open class CheckAbonentActions : Operation() {
 
     override val func: Scenario.() -> TransitionCondition = {
-        var transitionCondition: TransitionCondition = YES
+        this.functionBuilder {
+            var transitionCondition: TransitionCondition = YES
 
-        // decision imitation
-        val isActionAlreadyActive: Boolean = Random.nextBoolean()
-        // Do something
-        if (isActionAlreadyActive) {
-            transitionCondition = NO
+            // decision imitation
+            val isActionAlreadyActive: Boolean = Random.nextBoolean()
+            // Do something
+            if (isActionAlreadyActive) {
+                transitionCondition = NO
+            }
+            transitionCondition
         }
-        transitionCondition
     }
 }
 
 open class WriteOffMoney : Operation() {
 
     override val func: Scenario.() -> TransitionCondition = {
-        var transitionCondition: TransitionCondition = YES
+        this.functionBuilder {
+            var transitionCondition: TransitionCondition = YES
 
-        // decision imitation
-        val isThereEnoughMoney: Boolean = Random.nextBoolean()
-        // Do something
-        if (isThereEnoughMoney) {
-            transitionCondition = NO
+            // decision imitation
+            val isThereEnoughMoney: Boolean = Random.nextBoolean()
+            // Do something
+            if (isThereEnoughMoney) {
+                transitionCondition = NO
+            }
+            transitionCondition
         }
-        transitionCondition
+    }
+}
+
+open class DefaultErrorHandling : Operation() {
+
+    override val func: Scenario.() -> TransitionCondition = {
+        this.functionBuilder {
+            val error: String = this.getFromStorage("error")
+            println(error)
+            STOP_EXECUTION
+        }
+    }
+}
+
+open class SpecialErrorHandling : Operation() {
+
+    override val func: Scenario.() -> TransitionCondition = {
+        this.functionBuilder {
+            val error: String = this.getFromStorage("error")
+            println("Some super error handling: $error")
+            STOP_EXECUTION
+        }
     }
 }
