+++ nexign-kotlin-bp-dsl/src/main/kotlin/com/nexign/dsl/base/exceptions/Exceptions.kt
