+++ nexign-kotlin-bp-dsl/src/main/kotlin/com/nexign/dsl/base/specification/ErrorRootingMap.kt
@@ -1,0 +1,5 @@
+package com.nexign.dsl.base.specification
+
+import com.nexign.dsl.base.Operation
+
+class ErrorRootingMap : HashMap<Operation, Operation>()
